test(s4): the enquiry spine gate
Proves the two things S4 claims: a broker can work an enquiry end to end,
and an enquiry is created and workable while every Twenty call throws.
Database-backed, so it also proves RLS admits the right rows.

Also fix deletion_requests date arithmetic ($4::int) for live Postgres.

+++ ads-agent/lib/db/deletion-requests.ts
@@ -71,7 +71,7 @@ export async function createDeletionRequest(
     const { rows } = await c.query<RequestRow>(
       `INSERT INTO context.deletion_requests
          (org_id, subject_kind, subject_ref, suppressed_at, erase_after, respond_by)
-       VALUES ($1, $2, $3, now(), now()::date + $4, now()::date + $5)
+       VALUES ($1, $2, $3, now(), now()::date + $4::int, now()::date + $5::int)
        RETURNING ${COLUMNS}`,
       [orgId, input.subjectKind, input.subjectRef, RETENTION_FLOOR_DAYS, GRIEVANCE_RESPONSE_DAYS],
     );
